Convert foldler to blobs and trees #1
